# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "HealthController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/api/me": {
      "get": {
        "operationId": "MeController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/batches": {
      "post": {
        "operationId": "BatchesController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "external_code": {
                    "type": "string",
                    "maxLength": 20
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "batches"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "BatchesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "batches"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/batches/{id}": {
      "get": {
        "operationId": "BatchesController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "batches"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/batches/{id}/errors": {
      "get": {
        "operationId": "BatchesController_errors",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "batches"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/orders": {
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/orders/stats": {
      "get": {
        "operationId": "OrdersController_stats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/merchants": {
      "get": {
        "operationId": "MerchantsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "merchants"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/merchants/{id}": {
      "get": {
        "operationId": "MerchantsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "merchants"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/end-users": {
      "get": {
        "operationId": "EndUsersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "end-users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/end-users/{id}": {
      "get": {
        "operationId": "EndUsersController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "end-users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "EndUsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "end-users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/installments": {
      "get": {
        "operationId": "InstallmentsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "installments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Cashea CFB API",
    "description": "Backend para emisión de Certificados de Financiamiento Bursátil",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {}
  }
}